requirements.txt
cycler==0.10.0
kiwisolver==1.2.0
matplotlib==3.2.2
mysql-connector-python==8.0.20
numpy==1.19.0
pkg-resources==0.0.0
protobuf==3.12.2
pycairo==1.19.1
PyGObject==3.36.1
PyMySQL==0.9.3
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
